move WorkspaceResourcesService+Impl to workspaces/resources
